+++ cmd/jaeger/internal/integration/fixtures/pict_input_spans.txt
