+++ cmd/jaeger/internal/integration/receivers/storagereceiver/README.md
