+++ cmd/jaeger/internal/integration/receivers/storagereceiver/config.go
